+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `onExited` prop to `Modal`
+
 ## [8.64.0] - 2019-07-17
 
 ### Added